pyproject.toml
[project]
name = "comfyui-save-image-xmp"
version = "0.1.0"
description = "ComfyUI nodes for saving images with XMP metadata (PNG/WebP/JPEG)"
dependencies = ["Pillow", "tifffile"]
